Create percentage bar plots

Generate bar plots showing percentages or frequencies across groups. This function calculates percentages within groups and creates a stacked bar plot.

Usage

plot_percent(
  df,
  group.by,
  variable,
  facet.by = NULL,
  cols = NULL,
  legend.ncol = 1,
  position = "stack",
  bar.width = 0.85,
  bar.color = "white",
  orientation = "vertical",
  y.label = "Frequency (%)",
  x.label = "",
  theme = "line"
)

Arguments

df

Dataframe containing the data to plot.

group.by

Column name to group data by (x-axis).

variable

Column name for variable to show as different colors (fill).

facet.by

Optional column name to facet the plot by. Default is NULL.

cols

Optional vector of colors for fill. Default is NULL (uses default ggplot colors).

legend.ncol

Number of columns for the legend. Default is 1.

position

Position adjustment for bars. Options are "stack" (default) or "fill".

bar.width

Width of the bars. Default is 0.85.

bar.color

Color of the bar borders. Default is "white".

orientation

Plot orientation. Options are "vertical" (default) or "horizontal".

y.label

Label for the y-axis. Default is "Frequency (%)".

x.label

Label for the x-axis. Default is "" (empty).

theme

Theme to apply. Default is "line". Options are "line" or "border".

Value

A ggplot object

Examples

if (FALSE) { # \dontrun{
# Basic percentage plot
plot_percent(mtcars, group.by = "am", variable = "cyl")

# With custom colors
plot_percent(mtcars, 
             group.by = "am", 
             variable = "cyl", 
             cols = c("red", "blue", "green"))
             
# With faceting
plot_percent(mtcars, 
             group.by = "am", 
             variable = "cyl", 
             facet.by = "vs")
             
# Horizontal orientation
plot_percent(mtcars, 
             group.by = "am", 
             variable = "cyl", 
             orientation = "horizontal")
} # }